refactor: move TableAccessor to src/Framework/Database
