+++ src/Revenue/Repositories/Revenue.php
@@ -1,7 +1,7 @@
 <?php
 namespace Give\Revenue\Repositories;
 
-use Give\Framework\TableAccessor;
+use Give\Framework\Database\TableAccessor;
 
 /**
  * Class Revenue
